[hotfix] Move delta join  to common catalog and source ItCase.
